Refactor the format/lint logic into a simple API.
This moves much of the logic for performing a format/lint operation from
the command line tool and into a new module, "SwiftFormat", which is
intended to be used by anyone who wants to format or lint Swift code.
This will support integrating it into other tools in a better way than
shelling out to the command line.
