+++ Sources/swift-format/main.swift
@@ -12,128 +12,12 @@
 
 import Basic
 import Foundation
+import SwiftFormat
 import SwiftFormatConfiguration
 import SwiftFormatCore
-import SwiftFormatPrettyPrint
 import Utility
 
-fileprivate enum Mode: String, Codable, ArgumentKind {
-  case format
-  case lint
-  case dumpConfiguration = "dump-configuration"
-  case version
-
-  static var completion: ShellCompletion {
-    return .values([
-      ("format", "Format the provided files."),
-      ("lint", "Lint the provided files."),
-      ("dump-configuration", "Dump the default configuration as JSON to standard output."),
-    ])
-  }
-
-  /// Indicates whether the mode requires at least one input file passed as a positional argument.
-  var requiresFiles: Bool {
-    switch self {
-    case .format, .lint: return true
-    case .dumpConfiguration, .version: return false
-    }
-  }
-
-  init(argument: String) throws {
-    guard let mode = Mode(rawValue: argument) else {
-      throw ArgumentParserError.invalidValue(argument: argument, error: .unknown(value: argument))
-    }
-    self = mode
-  }
-}
-
-fileprivate struct CommandLineOptions {
-  var configurationPath: String? = nil
-  var paths: [String] = []
-  var verboseLevel = 0
-  var mode: Mode = .format
-  var debugOptions: DebugOptions = []
-}
-
-fileprivate func processArguments(
-  commandName: String, _ arguments: [String]
-) -> CommandLineOptions {
-  let parser = ArgumentParser(commandName: commandName,
-                              usage: "[options] <filename or path> ...",
-                              overview: "Format or lint Swift source code.")
-  let binder = ArgumentBinder<CommandLineOptions>()
-  binder.bind(
-    option: parser.add(
-      option: "--mode",
-      shortName: "-m",
-      kind: Mode.self,
-      usage: "The mode to run swift-format in. Either 'format', 'lint', or 'dump-configuration'."
-  )) {
-    $0.mode = $1
-  }
-  binder.bind(
-    option: parser.add(
-      option: "--version",
-      shortName: "-v",
-      kind: Bool.self,
-      usage: "Prints the version and exists"
-  )) { opts, _ in
-    opts.mode = .version
-  }
-  binder.bindArray(
-    positional: parser.add(
-      positional: "filenames or paths",
-      kind: [String].self,
-      optional: true,
-      strategy: .upToNextOption,
-      usage: "One or more input filenames",
-      completion: .filename
-  )) {
-    $0.paths = $1
-  }
-  binder.bind(
-    option: parser.add(
-      option: "--configuration",
-      kind: String.self,
-      usage: "The path to a JSON file containing the configuration of the linter/formatter."
-  )) {
-    $0.configurationPath = $1
-  }
-
-  // Add advanced debug/developer options. These intentionally have no usage strings, which omits
-  // them from the `--help` screen to avoid noise for the general user.
-  binder.bind(
-    option: parser.add(
-      option: "--debug-disable-pretty-print",
-      kind: Bool.self
-  )) {
-    $0.debugOptions.set(.disablePrettyPrint, enabled: $1)
-  }
-  binder.bind(
-    option: parser.add(
-      option: "--debug-dump-token-stream",
-      kind: Bool.self
-  )) {
-    $0.debugOptions.set(.dumpTokenStream, enabled: $1)
-  }
-
-  var opts = CommandLineOptions()
-  do {
-    let args = try parser.parse(arguments)
-    binder.fill(args, into: &opts)
-
-    if opts.mode.requiresFiles && opts.paths.isEmpty {
-      throw ArgumentParserError.expectedArguments(parser, ["filenames or paths"])
-    }
-  } catch {
-    stderrStream.write("error: \(error)\n\n")
-    parser.printUsage(on: stderrStream)
-    exit(1)
-  }
-  return opts
-}
-
-func main(_ arguments: [String]) -> Int32 {
+fileprivate func main(_ arguments: [String]) -> Int32 {
   let url = URL(fileURLWithPath: arguments.first!)
   let options = processArguments(commandName: url.lastPathComponent, Array(arguments.dropFirst()))
   switch options.mode {
